Add `filter-version` logic
