+++ src/sync.rs
@@ -22,6 +22,20 @@ impl From<anyhow::Error> for RustcPullError {
     }
 }
 
+#[derive(Copy, Clone)]
+pub enum FilterVersion {
+    /// Keep empty merge commits.
+    Version1,
+    /// Skip empty merge commits.
+    Version2,
+}
+
+impl FilterVersion {
+    pub fn latest() -> Self {
+        Self::Version2
+    }
+}
+
 pub struct PullResult {
     pub merge_commit_message: String,
 }
@@ -77,7 +91,7 @@ impl GitSync {
         let josh_url = josh.git_url(
             &upstream_repo,
             Some(&upstream_sha),
-            &self.context.config.construct_josh_filter(),
+            &construct_josh_filter(&self.context.config),
         );
 
         let orig_head = get_current_head_sha(self.verbose)?;
@@ -268,7 +282,7 @@ After you fix the conflicts, `git add` the changes and run `git merge --continue
         let josh_url = josh.git_url(
             &format!("{username}/rust"),
             None,
-            &self.context.config.construct_josh_filter(),
+            &construct_josh_filter(&self.context.config),
         );
         let user_upstream_url = format!("https://github.com/{username}/rust");
 
@@ -473,5 +487,160 @@ impl Drop for GitResetOnDrop {
             run_command(&["git", "reset", "--hard", &self.reset_to], self.verbose)
                 .expect(&format!("cannot reset current branch to {}", self.reset_to));
         }
+    }
+}
+
+fn construct_josh_filter(config: &JoshConfig) -> String {
+    let filter = match (&config.path, &config.filter) {
+        (Some(path), None) => format!(":/{path}"),
+        (None, Some(filter)) => filter.clone(),
+        _ => panic!("Config contains both path and a filter"),
+    };
+    match config.filter_version {
+        // Keep backwards compatibility with repositories that started with a legacy version of
+        // Josh.
+        FilterVersion::Version1 => {
+            // Convert old :rev syntax
+            let filter = convert_rev_syntax(&filter);
+            // Keep empty merges
+            wrap_compat(&filter)
+        }
+        // Use the current default behavior of Josh.
+        FilterVersion::Version2 => filter,
+    }
+}
+
+/// Converts filters from old `:rev(sha:filter)` syntax to new
+/// `:rev(<=sha:filter)` syntax. Null SHAs (40 zeros) become `_`.
+/// Only touches SHAs inside `:rev(...)` blocks.
+fn convert_rev_syntax(input: &str) -> String {
+    let rev_block = regex::Regex::new(r":rev\([^)]*\)").unwrap();
+    let entry = regex::Regex::new(
+        r"(?x)
+        ([,(])                # delimiter before entry
+        (0{40}|[0-9a-f]{40})  # full SHA
+        :                     # colon separator
+    ",
+    )
+    .unwrap();
+
+    rev_block
+        .replace_all(input, |block: &regex::Captures| {
+            entry
+                .replace_all(&block[0], |caps: &regex::Captures| {
+                    let delim = &caps[1];
+                    let sha = &caps[2];
+                    if sha.chars().all(|c| c == '0') {
+                        format!("{delim}_:")
+                    } else {
+                        format!("{delim}<={sha}:")
+                    }
+                })
+                .into_owned()
+        })
+        .into_owned()
+}
+
+/// Wraps a filter with the backwards compatibility meta options for
+/// trivial merge preservation and CRLF normalization in gpgsig headers.
+///
+/// `:your/filter` becomes
+/// `:~(history="keep-trivial-merges",gpgsig="norm-lf")[:your/filter]`
+fn wrap_compat(filter: &str) -> String {
+    format!(":~(history=\"keep-trivial-merges\",gpgsig=\"norm-lf\")[{filter}]")
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn no_rev_block_unchanged() {
+        assert_eq!(convert_rev_syntax(":/some/path"), ":/some/path");
+    }
+
+    #[test]
+    fn single_sha_gets_prefix() {
+        assert_eq!(
+            convert_rev_syntax(":rev(3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/some/path)"),
+            ":rev(<=3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/some/path)",
+        );
+    }
+
+    #[test]
+    fn null_sha_becomes_underscore() {
+        assert_eq!(
+            convert_rev_syntax(":rev(0000000000000000000000000000000000000000:/some/path)"),
+            ":rev(_:/some/path)",
+        );
+    }
+
+    #[test]
+    fn multiple_entries_in_rev_block() {
+        assert_eq!(
+            convert_rev_syntax(
+                ":rev(3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/p1,\
+                 e4c7a2d8f1b3e5a9d6c0f2b4a7e1d3c5f8a0b6e9:/p2,\
+                 0000000000000000000000000000000000000000:/p3)"
+            ),
+            ":rev(<=3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/p1,\
+             <=e4c7a2d8f1b3e5a9d6c0f2b4a7e1d3c5f8a0b6e9:/p2,\
+             _:/p3)",
+        );
+    }
+
+    #[test]
+    fn already_converted_syntax_unchanged() {
+        assert_eq!(
+            convert_rev_syntax(":rev(<=3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/some/path)"),
+            ":rev(<=3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/some/path)",
+        );
+    }
+
+    #[test]
+    fn underscore_syntax_unchanged() {
+        assert_eq!(
+            convert_rev_syntax(":rev(_:/some/path)"),
+            ":rev(_:/some/path)",
+        );
+    }
+
+    #[test]
+    fn sha_outside_rev_block_unchanged() {
+        assert_eq!(
+            convert_rev_syntax("3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/some/path"),
+            "3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/some/path",
+        );
+    }
+
+    #[test]
+    fn wrap_compat_simple_filter() {
+        assert_eq!(
+            wrap_compat(":/some/path"),
+            ":~(history=\"keep-trivial-merges\",gpgsig=\"norm-lf\")[:/some/path]",
+        );
+    }
+
+    #[test]
+    fn wrap_compat_rev_filter() {
+        assert_eq!(
+            wrap_compat(
+                ":rev(75dd959a3a40eb5b4574f8d2e23aa6efbeb33573:prefix=src/tools/miri):/src/tools/miri"
+            ),
+            ":~(history=\"keep-trivial-merges\",gpgsig=\"norm-lf\")\
+             [:rev(75dd959a3a40eb5b4574f8d2e23aa6efbeb33573:prefix=src/tools/miri):/src/tools/miri]",
+        );
+    }
+
+    #[test]
+    fn multiple_rev_blocks() {
+        assert_eq!(
+            convert_rev_syntax(
+                ":rev(3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/p1)\
+                 :rev(e4c7a2d8f1b3e5a9d6c0f2b4a7e1d3c5f8a0b6e9:/p2)"
+            ),
+            ":rev(<=3a1f5e2b9c8d4e7f6a0b1c2d3e4f5a6b7c8d9e0f:/p1)\
+             :rev(<=e4c7a2d8f1b3e5a9d6c0f2b4a7e1d3c5f8a0b6e9:/p2)",
+        );
     }
 }
